Apply review cleanups: dead code, API surface, shared test helpers
Drops the clientPools WeakMap left behind when the transitional pool
escape hatch was deleted, and the isEnabled prop that
RecorderHeartbeatListener declared but never used. The library's entry
point now exports only getObserverInfo among the side-mapping helpers —
the other five had no consumers outside the library and were needless
public API for a future protocol version to honor.

The emitTo / stationConnected / camera-controls store factory helpers
that had been copy-pasted across the new test files move to
tests/imaging-test-utils.ts beside the wire harness.
