+++ react-frontend/src/features/listeners/CamHeartbeatListener.test.tsx
@@ -1,43 +1,15 @@
 import { afterEach, expect, test, vi } from "vitest";
 import React from "react";
 import { cleanup } from "@testing-library/react";
-import { configureStore } from "@reduxjs/toolkit";
-import cameraControlsReducer from "../camera-controls/cameraControlsSlice.js";
 import { createSocketIoHarness } from "../../../tests/socket.io-harness";
+import {
+  emitTo,
+  makeCameraControlsStore,
+  stationConnected,
+} from "../../../tests/imaging-test-utils";
 import { renderWithProviders } from "../../../tests/renderWithProviders";
 import { getSharedImagingClient } from "../../lib/imaging-client";
 import CamHeartbeatListener from "./CamHeartbeatListener.jsx";
-
-type CameraControlsState = ReturnType<typeof cameraControlsReducer>;
-
-function makeStore(overrides: Partial<CameraControlsState> = {}) {
-  const baseState = cameraControlsReducer(undefined, { type: "@@INIT" } as any);
-  return configureStore({
-    reducer: { cameraControls: cameraControlsReducer },
-    preloadedState: { cameraControls: { ...baseState, ...overrides } },
-  });
-}
-
-/** Emit a server → client event on a specific Socket.IO namespace. */
-function emitTo(h: any, namespace: string, event: string, ...args: any[]) {
-  h.emit({ event, namespace }, ...args);
-}
-
-/** Resolves once the given side's namespace connection is established. */
-function stationConnected(side: string): Promise<void> {
-  return new Promise((resolve) => {
-    const unsubscribe = getSharedImagingClient()
-      .station(side)
-      .onConnectionStatus(({ status }) => {
-        if (status === "connected") {
-          queueMicrotask(() => {
-            unsubscribe();
-            resolve();
-          });
-        }
-      });
-  });
-}
 
 afterEach(() => {
   cleanup();
@@ -46,10 +18,10 @@ afterEach(() => {
 
 test("feeds the observer's own heartbeat into the main reducer", async () => {
   const h = createSocketIoHarness();
-  const store = makeStore({ observerSide: "P" });
+  const store = makeCameraControlsStore({ observerSide: "P" });
   renderWithProviders(<CamHeartbeatListener />, { store });
 
-  await stationConnected("P");
+  await stationConnected(getSharedImagingClient().station("P"));
   emitTo(h, "/port", "CamHeartbeat", {
     camera: "port_brow_4k",
     focus_mode: "MF",
@@ -70,13 +42,15 @@ test.each([
   "pilot override %s routes into %s",
   async (namespaceOverride, stateField) => {
     const h = createSocketIoHarness();
-    const store = makeStore({ observerSide: "PL" });
+    const store = makeCameraControlsStore({ observerSide: "PL" });
     renderWithProviders(
       <CamHeartbeatListener namespaceOverride={namespaceOverride} />,
       { store }
     );
 
-    await stationConnected(namespaceOverride);
+    await stationConnected(
+      getSharedImagingClient().station(namespaceOverride)
+    );
     emitTo(h, namespaceOverride, "CamHeartbeat", { camera: "some_cam" });
 
     await vi.waitFor(() =>
