+++ react-frontend/src/features/listeners/RecorderHeartbeatListener.test.tsx
@@ -1,41 +1,15 @@
 import { afterEach, expect, test, vi } from "vitest";
 import React from "react";
 import { cleanup } from "@testing-library/react";
-import { configureStore } from "@reduxjs/toolkit";
-import cameraControlsReducer from "../camera-controls/cameraControlsSlice.js";
 import { createSocketIoHarness } from "../../../tests/socket.io-harness";
+import {
+  emitTo,
+  makeCameraControlsStore,
+  stationConnected,
+} from "../../../tests/imaging-test-utils";
 import { renderWithProviders } from "../../../tests/renderWithProviders";
 import { getSharedImagingClient } from "../../lib/imaging-client";
 import RecorderHeartbeatListener from "./RecorderHeartbeatListener.jsx";
-
-type CameraControlsState = ReturnType<typeof cameraControlsReducer>;
-
-function makeStore(overrides: Partial<CameraControlsState> = {}) {
-  const baseState = cameraControlsReducer(undefined, { type: "@@INIT" } as any);
-  return configureStore({
-    reducer: { cameraControls: cameraControlsReducer },
-    preloadedState: { cameraControls: { ...baseState, ...overrides } },
-  });
-}
-
-function emitTo(h: any, namespace: string, event: string, ...args: any[]) {
-  h.emit({ event, namespace }, ...args);
-}
-
-function stationConnected(side: string): Promise<void> {
-  return new Promise((resolve) => {
-    const unsubscribe = getSharedImagingClient()
-      .station(side)
-      .onConnectionStatus(({ status }) => {
-        if (status === "connected") {
-          queueMicrotask(() => {
-            unsubscribe();
-            resolve();
-          });
-        }
-      });
-  });
-}
 
 afterEach(() => {
   cleanup();
@@ -44,10 +18,10 @@ afterEach(() => {
 
 test("stores recorder heartbeats for the observer's own side", async () => {
   const h = createSocketIoHarness();
-  const store = makeStore({ observerSide: "S" });
+  const store = makeCameraControlsStore({ observerSide: "S" });
   renderWithProviders(<RecorderHeartbeatListener />, { store });
 
-  await stationConnected("S");
+  await stationConnected(getSharedImagingClient().station("S"));
   emitTo(h, "/stbd", "RecorderHeartbeat", {
     command: "SRVS",
     camera: "Stbd Brow",
